Fixed many security issues displaying in the github security, added .env.examples files and update README.md file

+++ Backend/controllers/expenseController.js
@@ -1,6 +1,7 @@
 // Import necessary packages and models
 const Expense = require("../models/Expense");
 const ExcelJS = require("exceljs");
+const mongoose = require("mongoose");
 
 /**
  * @desc    Add a new expense
@@ -72,9 +73,13 @@ exports.deleteExpense = async (req, res) => {
   try {
     const userId = req.user.id;
     const expenseId = req.params.id;
+    if (!mongoose.Types.ObjectId.isValid(expenseId)) {
+      return res.status(400).json({ message: "Invalid expense id" });
+    }
+
     const expense = await Expense.findOneAndDelete({
-      _id: expenseId,
-      user: userId,
+      _id: mongoose.Types.ObjectId(expenseId),
+      user: mongoose.Types.ObjectId(userId),
     });
     if (!expense) {
       return res.status(404).json({ message: "Expense not found" });
@@ -101,9 +106,15 @@ exports.updateExpense = async (req, res) => {
     const userId = req.user.id;
     const expenseId = req.params.id;
     const { title, icon, amount, category, date, description } = req.body;
+    if (!mongoose.Types.ObjectId.isValid(expenseId)) {
+      return res.status(400).json({ message: "Invalid expense id" });
+    }
+
+    const updatePayload = { title, icon, amount, category, date, description };
+
     const expense = await Expense.findOneAndUpdate(
-      { _id: expenseId, user: userId },
-      { title, icon, amount, category, date, description },
+      { _id: mongoose.Types.ObjectId(expenseId), user: mongoose.Types.ObjectId(userId) },
+      updatePayload,
       { new: true, runValidators: true },
     );
     if (!expense) {
